Move debounce hook to global utils dir
